Adds Content-Length header in simple app
The simple app in the integration test suite was accidentally emitting a
response that used chunked transfer encoding. This resulted in response
bodies that had strange extra characters. This is the default behavior
for Rack when the response headers do not include a Content-Length
header. This Stack Overflow post explains the issue a bit more:
https://stackoverflow.com/questions/23128957/strange-extra-characters-in-rails-response-body.

Resolves #75